templates/django_backend/requirements.txt
asgiref==3.8.1
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.0.7
django-allauth==65.7.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
gunicorn==23.0.0
hashids==1.3.1
idna==3.11
inflection==0.5.1
packaging==25.0
psycopg2==2.9.10
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.16.0
sqlparse==0.5.1
urllib3==2.5.0
